Added Dynamic Gamepad Detection plugin + Guide
Added Dynamic Gamepad Detection plugin + Guide

Detects every frame for connection/disconnection of gamepad. Probably should be optimised in the future to only check every 5 seconds or so I guess but it's not a major issue for small games.

Plus the faster it checks the more consistent it will be to not lose input when re-connection occurs after changing batteries in disposeable battery gamepads like xbox, useful for more action related games, or pausing/unpausing based on controller connection.