Added plugin for Changing Map Names
Added DynamicMapNames to allow changing of map names at runtime. Unsure at this stage if the names persist through saving, but it's a start.
Use case: Eg. map name is empty or a vague descriptive name of the map is provided in the rpg maker MV editor for an "Unknown place" that the character has no info about.
Player talks to NPC that provides a name to the place and updates the map name (only updates once reloading the map exit and re-enter)

ChangeMapName mapID new_map_name | Changes the name of the mapID to the new string provided.
ClearMapName mapID | Clears the custom name back to the original one set in the editor.