Added a Dialogue Command
Added a very simple dialogue command that can be used for voice acting. At this stage support for voice skipping is not added. However, it is a start.
Allows developers to add an inline command to their show text " \voice[filename] " to play sound effect files with voices of the lines being read.

Pros:
Helps with better timing of voices or sound effects (eg. an explosion when the word "explosion" is about to be printed).
Condenses event structure inside the editor.